Extract keyDownDelegate logic into keyboard-shortcuts.js
Moves the ~55-line conditional body of keyDownDelegate out of
event-listeners-add.js and into a dedicated module, so the delegate
stays thin (two lines) and consistent with clickDelegate/changeDelegate.
No behaviour changes.
